Not overwrite klayoutrc.
Not overwrite klayoutrc.

+++ pdk_OpenSUSI-TR10-setup.sh
@@ -125,12 +125,15 @@ fi
 if [ ! -d "$HOME/.klayout/salt/$PDK/libraries/" ]; then
 	mkdir -p "$HOME/.klayout/salt/$PDK/libraries/"
 fi
+
+if [ ! -f "$HOME/.klayout/klayoutrc" ]; then
+	cp -f $SRC_DIR/TR-1um/libs.tech/klayout/klayoutrc $HOME/.klayout/
+fi
 cp -aR $SRC_DIR/TR-1um/libs.tech/klayout/tech/* $HOME/.klayout/salt/$PDK/
 mv     $HOME/.klayout/salt/$PDK/TR-1um.lyp $HOME/.klayout/salt/$PDK/tech/
 mv     $HOME/.klayout/salt/$PDK/TR-1um.lyt $HOME/.klayout/salt/$PDK/tech/
 cp -aR $SRC_DIR/TR-1um/libs.tech/klayout/libraries/* $HOME/.klayout/salt/$PDK/libraries/
 cp -aR $SRC_DIR/TR-1um/libs.tech/klayout/grain.xml $HOME/.klayout/salt/$PDK/
-cp -f $SRC_DIR/TR-1um/libs.tech/klayout/klayoutrc $HOME/.klayout/
 rm -fr $HOME/.klayout/salt/$PDK/drc/run_IP62.drc
 rm -fr $HOME/.klayout/salt/$PDK/drc/drc_IP62.lydrc
 rm -fr $HOME/.klayout/salt/$PDK/drc/IP62/
